푼 문제를 search 명령어 사용 시 색깔로 표시하는 기능 추가
